vision-service/requirements.txt
opencv-python
ultralytics
kafka-python
numpy